docker: fix live restore
There is a known issue when trying to delete a container.  Since the
shm mount point can be busy in other mount namespaces (the shim
processes for the dockerd containers), the rm -rf for the container
might fail.  As a workaround delete, try to delete them when the
service is started again.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #8
Approved by: giuseppe