docker-fedora: use a keep alive process to re-join the same namespace
Using a shared rootfsPropagation gives a lot of troubles and leak
mounts.  Just use a running keep alive process to join on a restart of
the Docker container.

In addition, the previous version was not working at all in case of a
Docker container update, as the rootfs of the container was being
deleted.  Supporting updates for the Docker engine is the main reason
of having live restore.

This MUST still be considered a proof of concept, not ready for any
usage yet.

Another alternative is to not use a separate mount namespace and only
setup the chroot environment where Docker runs, but this brings
another set of problems with ensuring the mount points are properly
cleaned up and not leaked.

The mount-in-ns tool is based on:

https://lists.projectatomic.io/projectatomic-archives/atomic-devel/2015-February/msg00064.html

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>

Closes: #22
Approved by: rhatdan