Clarify docker and container-engine
As discussed, the docker system container should still be called
docker to match the package. The container-engine system container,
which was designed as an abstract engine, will be split out. Taking
advantage of the Fedora layered build system, it will currently
just layer a README on top of the docker image (since we only support
docker at the moment). In the future, if container-engine gets
expanded, it will be easy to make changes just to the container-engine
container.

Signed-off-by: Yu Qi Zhang <jerzhang@redhat.com>

Closes: #75
Approved by: giuseppe