Create cri-o system container
Create a system container for cri-o, built from source inside the
container. Currently this is still experimental, but has been
tested to work on f25/26. In the future the build script will
likely be replaced by the packaged version.

The container image comes with crioctl installed on the host, but
does not include cni configs. The daemon will run, but to create
pods/containers, the user must still set up cni configs on the host.

Signed-off-by: Yu Qi Zhang <jerzhang@redhat.com>