cri-o-fedora: prevent a bind mounts bomb
Under RHELAH I've seen that when cri-o is restarted a few times, we
create a lot of mount points.  That is because we are bind mounting as
rshared both /var/lib and /var/lib/origin.  We needed /var/lib because
we want to share bind mounts from /var/lib/containers.  Solve it by
creating a new bind mount and make /var/lib private.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>