crio-o-centos: drop host mount /var/lib
Having a rbind bind mount on /var/lib is causing all sort of issues in
the host (e.g. ostree storage on two different file systems), so drop
it completely and in case we will need to create mounts in the host
under /var/lib, we will deal with that more specifically.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>