Enable mount propagation in /var/lib/kubelet
Cases where an external volume driver manages the bind mounts to be made
available to containers in a different namespace need mount propagation
on /var/lib/kubelet. An example would be a driver relying on FlexVolume
which is doing the pod bind mount to a shared filesystem.

Closes: #124
Approved by: jasonbrooks